fix(apps): audit round 1 — collapse the module split, resolve the glyph per branch
Adversarial audit findings on #3540.

- Collapse `appListingActionGlyphIcons.ts` back into
  `appListingActionGlyph.ts`. The split's stated rationale — "keep it
  node-importable" — was falsified by this PR's own test, which imports the
  tabler module in the node project and passes in 4ms. The cited precedent
  (`marketplaceCategoryIcons.ts`) is a SINGLE module carrying its own tabler
  imports, so the split was a new convention, not the existing one.

- Resolve the glyph INSIDE each branch of `PrimaryAction`, from the mode that
  branch has established, instead of once up front from `action.mode`.
  `href` is optional on `DetailPrimaryAction`, so an `open`/`visit` action
  without one falls through to the informational return — and the hoisted
  lookup would have painted a launch icon there. Not reachable from today's
  `getDetailPrimaryAction` (every `open`/`visit` return sets an href), so this
  was latent, not live; the point is that the pre-change code was correct by
  construction and the hoisted version was correct only by coincidence of the
  current view model. The fallthrough now hard-codes the `info` glyph.

- Test comment claimed `readonly DetailActionMode[]` makes the mode list
  exhaustive. It does not — that type accepts any subset, so a 5th mode would
  typecheck and silently under-cover the distinctness assertions. Replaced with
  `satisfies Record<DetailActionMode, true>`, which genuinely fails to compile
  on a missing key. Same for the card actions.

- Removed `resolves an icon for every glyph`: it cannot fail while the file
  typechecks, since `Record<PrimaryActionGlyph, Icon>` already forbids a
  missing key. It read as coverage while asserting nothing. Replaced with a
  real distinctness assertion over the card actions.

- `renderAndSettle` timeout 5000 -> 10000, matching the project-wide
  `vi.waitFor` default that `test/component-setup.tsx` sets deliberately for
  the saturated preview CI box. `vi.waitUntil` is not covered by that patch.

Re-verified after the restructure: mutating the `open` branch to
`glyphFor('visit')` still fails the component gate with
`expected 'external-link' not to be 'external-link'` (17 pass, 1 fail).
Node 9/9, component 18/18, tsc 22 pre-existing errors (none in these files),
eslint 0 errors.
