+++ src/components/Apps/appListingActionGlyph.ts
@@ -1,9 +1,18 @@
+import {
+  IconExternalLink,
+  IconInfoCircle,
+  IconPlayerPlay,
+  IconPlugConnected,
+} from '@tabler/icons-react';
+import type { Icon } from '@tabler/icons-react';
 import type { ListingCtaAction } from '~/components/Apps/appListingCardView';
 import type { DetailActionMode } from '~/components/Apps/appListingDetailView';
 
 /**
  * The primary-action GLYPH vocabulary — the single source of truth for "which
  * icon does this CTA wear", shared by the store card and the listing detail.
+ * Same shape as `marketplaceCategoryIcons.ts`: one module carrying both the
+ * mapping logic and the icon record it resolves to.
  *
  * WHY THIS EXISTS. civitai #3391 removed the kind + category badges from both
  * `AppListingCard` and `AppListingDetailBody`, on the stated grounds that the
@@ -17,16 +26,29 @@ import type { DetailActionMode } from '~/components/Apps/appListingDetailView';
  * Both view-models ALREADY carry the discriminator (`DetailActionMode` /
  * `ListingCtaAction`); the components threw it away at the icon. So the fix is a
  * mapping, and the invariant worth pinning is that the in-site and off-site
- * glyphs are DIFFERENT — which is a pure function of the mode, and therefore
- * testable in the plain node project rather than by asserting rendered SVG.
- *
- * 🔴 KEPT RUNTIME-PURE ON PURPOSE — type-only imports, no React, no
- * `@tabler/icons-react`. The glyph→icon record lives next door in
- * `appListingActionGlyphIcons.ts` (the `marketplaceCategoryIcons.ts` pattern),
- * so this module stays importable from a node unit test. Do not add a value
- * import here.
+ * glyphs are DIFFERENT.
  */
 export type PrimaryActionGlyph = 'launch' | 'external' | 'connect' | 'info';
+
+/**
+ * Glyph → Tabler icon.
+ *
+ * 🔴 `launch` and `external` must stay DIFFERENT icons. The whole point of the
+ * vocabulary is that "runs here" and "leaves the site" look different — mapping
+ * them to the same icon reinstates exactly the defect #3391's badge removal
+ * assumed was already handled. Pinned in
+ * `__tests__/appListingActionGlyph.test.ts`.
+ */
+export const ACTION_GLYPH_ICONS: Record<PrimaryActionGlyph, Icon> = {
+  /** In-site nav to the in-host page runner — a play/run affordance, no new tab. */
+  launch: IconPlayerPlay,
+  /** Leaves civitai.com in a new tab (`target="_blank"` + `rel="noopener noreferrer"`). */
+  external: IconExternalLink,
+  /** OAuth connect affordance (stubbed until the cutover wires it). */
+  connect: IconPlugConnected,
+  /** Informational — no launch, no navigation off-site. */
+  info: IconInfoCircle,
+};
 
 /**
  * Detail-page primary action → glyph.
@@ -34,6 +56,14 @@ export type PrimaryActionGlyph = 'launch' | 'external' | 'connect' | 'info';
  * `open` is in-site nav to `/apps/run/<slug>` (no `target`); `visit` is the
  * external new-tab anchor. They MUST NOT share a glyph — that equality is the
  * defect this module exists to prevent regressing.
+ *
+ * 🔴 Resolve this from the mode the RENDERING BRANCH has established, not from a
+ * mode a later guard may still reject. `DetailPrimaryAction.href` is optional,
+ * so `PrimaryAction`'s `mode === 'open' && action.href` branch can in principle
+ * be skipped while the mode is still `'open'` — resolving the glyph once, up
+ * front, would then paint a launch icon on the informational fallback. Not
+ * reachable from today's `getDetailPrimaryAction` (every `open`/`visit` return
+ * sets an href), which is precisely why it would be easy to introduce later.
  */
 export function detailActionGlyph(mode: DetailActionMode): PrimaryActionGlyph {
   switch (mode) {
